Code Reorganisation
- Move three functions from `bitcoin.py` to `util.py` as a temporary
  measure.

- Moved the rest of `bitcoin.py` to `transaction.py`.

`bitcoin` is a terrible name for a module (for obvious reasons), and
most of what was contained in it was related to transaction construction
and serialisation. This move, however, is incomplete, as some of the
code that is now in `transaction.py` should actually be elsewhere.

Updated test suite to deal with new module name.