Be Able to Run Without a Backend, using `--force`.
If `--force` is used, don’t check that the backend is available, and
when polling for blocks in `blocks.follow()`, if the backend can’t be
reached, just sleep and try again repeatedly.

This is useful for testing purposes, when a complete copy of the
blockchain isn’t available, but a reasonably complete database is.